IXWW22: display: Add TFT ASIL and Connected feature
Add support of detecting TFT_ASIL and TFT_CONNECTED.
For ASIL, if the signal is being sensed as abnormal, we will reset
the system.
For CONNECTED, we will check on initialization procedure whether
the display is connected, if not, we will not configure related
pins.

Testing and Verification Instruction:
- Disconnect the display and power on, check if the display related
  signal will not be configured.
- Force trigger asil signal, check if the system will reset itself.

Type: New Feature
Issue: IXWW22-1183
Change-Id: I51ea88f08605c8cd5feafa926155702ae6c0fcc0
Reviewed-on: https://gerrit.garmin.com.tw/code-review/c/autooem/products/nxp/rt1172/+/75451
Reviewed-by: Huang, Kenny <kenny.huang@garmin.com>
Reviewed-by: Kuo, Dilbert <dilbert.kuo@garmin.com>
Tested-by: Huang, Kenny <kenny.huang@garmin.com>
Tested-by: Kuo, Dilbert <dilbert.kuo@garmin.com>
Approve-Change: Kuo, Dilbert <dilbert.kuo@garmin.com>