IXWW22: Fix TFT_BL_PWM frequency not correct
We need to modify the root clock of FLEXIO2 for correct frequency.
But currently if we set to 20k, the result is 14k.

Temporarily fine tune this value to 27k so the result is 20k.
This needs to be fix if the clock tool is available

(cherry picked from commit 20e00d9af0cfa62ee3558bbea62cf39a6aeb12db)

Tpye: Bug Fix
Issue: AUTOOEM-IXWW22-4614
Change-Id: I08ff6145016270553f40aeb740ad8584f36c8118
Reviewed-on: https://gerrit.garmin.com.tw/code-review/c/autooem/products/nxp/rt1172/+/80135
Reviewed-by: Huang, Kenny <kenny.huang@garmin.com>
Tested-by: Huang, Kenny <kenny.huang@garmin.com>
Tested-by: Liao, Edward <edward.liao@garmin.com>
Approve-Change: Liao, Edward <edward.liao@garmin.com>