Fix time sync issue when changing playback rate
It's possible for user to change the playback rate while listening
the program such as podcast or audiobook. However, elapsed time and
seek bar remain unchanged on the media UI even though playback rate
is changed.

Root cause is that we start/stop timer based on play/pause state only.
Therefore, in current design, when playback rate is changed and then
the fast forwarding status is sent to LC, timer is stopped so that
the time and seek bar is not synchronized with the media app.

In this patch, fast forwarding check and rewinding check are added to
ensure the elapsed time and seek bar are updated when receiving
fast forwarding/rewinding status.

Testing and Verification Instructions:
  1. Build and install software.
  2. Connect BT and BLE.
  3. Search and play the podcast on the spotify.
  4. Change playback rate to other number (default is 1.0, make it
     smaller or greater than 1.0).
  5. Check if elapsed time and seek bar are synchronized with
     media app.

Issue: Autooem-IXWW22-1583
Issue: Autooem-IXWW22-1586
Change-Id: I23a96c2b29e83d10361d56deb75730488d4043f8
Type: Bug Fix
Reviewed-on: https://gerrit.garmin.com.tw/code-review/c/autooem/products/nxp/rt1172/+/83202
Reviewed-by: Liao, Edward <edward.liao@garmin.com>
Tested-by: Liao, Edward <edward.liao@garmin.com>
Approve-Change: Liao, Edward <edward.liao@garmin.com>