Fix value conversion for eco visualizer
Root cause:
I misunderstood the value conversion rule on the spec so that the
relative value is not set to max when H'000 is received if current
unit is L/100km.

Countermeasure
Set relative value to max when H'000 is received if current unit is
L/100km.

Testing and Verification Instructions:
  1. Send Fuel unit to L/100km
     00EB001A010300000000000000
  2. Enter eco visualizer, set 00EB00080200FF000000000000 to engine
     speed and see if animation is displayed.
  3. Set 00EB0027020190000000000000 to fuel average and set
     00EB00290201FF000000000000 to fuel instant.
  4. Set 00EB0029020000000000000000 to fuel instant , check if
     entire eco bar is displayed(maximum level).

Issue: Autooem-IXWW22-6459
Change-Id: Ibf124dbe5c51677b728c67079f5f6d990773143a
Type: Bug Fix
Reviewed-on: https://gerrit.garmin.com.tw/code-review/c/autooem/products/nxp/rt1172/+/88725
Reviewed-by: Liao, Edward <edward.liao@garmin.com>
Tested-by: Liao, Edward <edward.liao@garmin.com>
Approve-Change: Liao, Edward <edward.liao@garmin.com>